Location is no longer stored if it is invalid
Previously the app would save a users location to re-render the page for them on a refresh, regardless of tha validity of the location. Now the location is only saved when the location is valid. Updated a test to represent that change, because it is now running in the fetch call and catch, which doesn't translate to testing